# Reduce the number of namespace version lookups in the *Multiple methods