Currently the namespace version is pulled from the cache storage once for every id sent to the `*Multiple()` methods in the `CacheProvider`. This request aim to reduce the amount of these lookups by introducing a new method `prepareNamespacedIdsMap()` which pulls the version number only once for the whole batch of ids.