@rado-h 2018-11-11
@lcobucci, to be honest, I completely missed that  the number is already cached as a class property. As it is the patch is just a micro optimization by reducing the number of function calls - the main benefits come in the fetch method where `array_combine($keys, array_map([$this, 'getNamespacedId'], $keys))` is flattened to a single array traversal in `prepareNamespacedIdsMap()`. After looking at the benchmark results I'm not sure if this patch offers enough gains :/

Here are the phpbench results before and after the patch (executed with php7.2.10)
```php
$cache = new \Doctrine\Common\Cache\ArrayCache();
$cache->fetchMultiple(range(1, 10000));
```
```
#BEFORE 
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 21,689.300 [22,118.540 22,188.684] 22,494.300 (μs)
⅀T: 110,592.700μs μSD/r 270.140μs μRSD/r: 1.221%
suite: 133f077ad5d7ea77fdf544df739811556f7f4fa7, date: 2018-11-11, stime: 19:27:09
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak   | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 5,365,504b | 22,215.300μs | +0.36σ       | +0.44%         |
| TestBench | benchMulti |        | []     | 10   | 1    | 5,365,504b | 21,689.300μs | -1.59σ       | -1.94%         |
| TestBench | benchMulti |        | []     | 10   | 2    | 5,365,504b | 21,976.200μs | -0.53σ       | -0.64%         |
| TestBench | benchMulti |        | []     | 10   | 3    | 5,365,504b | 22,217.600μs | +0.37σ       | +0.45%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 5,365,504b | 22,494.300μs | +1.39σ       | +1.70%         |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
#AFTER
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 19,850.900 [19,992.820 19,915.895] 20,189.800 (μs)
⅀T: 99,964.100μs μSD/r 132.522μs μRSD/r: 0.663%
suite: 133f077e5eb533e2062dedd54eb1135575fde094, date: 2018-11-11, stime: 19:27:43
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak   | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 4,841,424b | 19,972.000μs | -0.16σ       | -0.1%          |
| TestBench | benchMulti |        | []     | 10   | 1    | 4,841,424b | 20,093.700μs | +0.76σ       | +0.50%         |
| TestBench | benchMulti |        | []     | 10   | 2    | 4,841,424b | 19,857.700μs | -1.02σ       | -0.68%         |
| TestBench | benchMulti |        | []     | 10   | 3    | 4,841,424b | 19,850.900μs | -1.07σ       | -0.71%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 4,841,424b | 20,189.800μs | +1.49σ       | +0.99%         |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
```




```php
$cache = new \Doctrine\Common\Cache\ArrayCache();
$cache->saveMultiple(range(1, 10000));
```
```
#BEFORE
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 22,321.600 [22,937.440 23,141.212] 23,450.500 (μs)
⅀T: 114,687.200μs μSD/r 403.501μs μRSD/r: 1.759%
suite: 133f077e3e9a1b1a888ff08473d6d634d901ae2e, date: 2018-11-11, stime: 19:28:11
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak   | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 9,656,576b | 23,099.400μs | +0.40σ       | +0.71%         |
| TestBench | benchMulti |        | []     | 10   | 1    | 9,656,576b | 22,639.900μs | -0.74σ       | -1.3%          |
| TestBench | benchMulti |        | []     | 10   | 2    | 9,656,576b | 23,175.800μs | +0.59σ       | +1.04%         |
| TestBench | benchMulti |        | []     | 10   | 3    | 9,656,576b | 23,450.500μs | +1.27σ       | +2.24%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 9,656,576b | 22,321.600μs | -1.53σ       | -2.68%         |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+

#AFTER
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 21,388.400 [21,691.180 21,625.446] 22,127.000 (μs)
⅀T: 108,455.900μs μSD/r 272.998μs μRSD/r: 1.259%
suite: 133f07795aa6a280389882b3b87e500e6e8e1608, date: 2018-11-11, stime: 19:28:42
+-----------+------------+--------+--------+------+------+-------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak    | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+-------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 10,189,392b | 21,760.300μs | +0.25σ       | +0.32%         |
| TestBench | benchMulti |        | []     | 10   | 1    | 10,189,392b | 21,771.500μs | +0.29σ       | +0.37%         |
| TestBench | benchMulti |        | []     | 10   | 2    | 10,189,392b | 21,408.700μs | -1.03σ       | -1.3%          |
| TestBench | benchMulti |        | []     | 10   | 3    | 10,189,392b | 22,127.000μs | +1.60σ       | +2.01%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 10,189,392b | 21,388.400μs | -1.11σ       | -1.4%          |
+-----------+------------+--------+--------+------+------+-------------+--------------+--------------+----------------+
```



```php
$cache = new \Doctrine\Common\Cache\ArrayCache();
$cache->deleteMultiple(range(1, 10000));
```

```
#BEFORE
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 13,047.600 [13,139.160 13,175.544] 13,219.200 (μs)
⅀T: 65,695.800μs μSD/r 64.338μs μRSD/r: 0.490%
suite: 133f0770e9177add2002955ce4ac0cdd4f6222a6, date: 2018-11-11, stime: 19:29:35
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak   | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 4,837,504b | 13,155.500μs | +0.25σ       | +0.12%         |
| TestBench | benchMulti |        | []     | 10   | 1    | 4,837,504b | 13,189.800μs | +0.79σ       | +0.39%         |
| TestBench | benchMulti |        | []     | 10   | 2    | 4,837,504b | 13,047.600μs | -1.42σ       | -0.7%          |
| TestBench | benchMulti |        | []     | 10   | 3    | 4,837,504b | 13,219.200μs | +1.24σ       | +0.61%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 4,837,504b | 13,083.700μs | -0.86σ       | -0.42%         |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+

#AFTER
1 subjects, 5 iterations, 10 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 11,529.700 [11,733.080 11,657.755] 12,075.000 (μs)
⅀T: 58,665.400μs μSD/r 182.897μs μRSD/r: 1.559%
suite: 133f077add1d7757ce1e2f41d5b75509609d5b0d, date: 2018-11-11, stime: 19:30:08
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| benchmark | subject    | groups | params | revs | iter | mem_peak   | time_rev     | comp_z_value | comp_deviation |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
| TestBench | benchMulti |        | []     | 10   | 0    | 4,841,456b | 12,075.000μs | +1.87σ       | +2.91%         |
| TestBench | benchMulti |        | []     | 10   | 1    | 4,841,456b | 11,665.100μs | -0.37σ       | -0.58%         |
| TestBench | benchMulti |        | []     | 10   | 2    | 4,841,456b | 11,667.200μs | -0.36σ       | -0.56%         |
| TestBench | benchMulti |        | []     | 10   | 3    | 4,841,456b | 11,728.400μs | -0.03σ       | -0.04%         |
| TestBench | benchMulti |        | []     | 10   | 4    | 4,841,456b | 11,529.700μs | -1.11σ       | -1.73%         |
+-----------+------------+--------+--------+------+------+------------+--------------+--------------+----------------+
```